add emoji image generation tool
Initially, I tried walking the known emoji unicode ranges, and using CoreText to tell me if the font had a glyph for the code point, but that didn't work so I snagged a category from a GitHub project and iterate over that.

It appears as though drawing at 144pt is a bit larger than the font supports, but that can be looked at some other time.